Fix component structure
Forgot pages router can't use nested components in routes.
