+++ src/components/project/case-study/components/layout/index.tsx
