+++ src/components/project/case-study/index.tsx
