+++ src/pages/_app.tsx
@@ -4,7 +4,7 @@ import { AnimatePresence } from "framer-motion";
 import Navigation from "@/components/navigation";
 import { VisitProvider } from "@/contexts/VisitContext";
 import Layout from "@/components/layout";
-import { useBaseRoute } from "@/hooks/use-base-route";
+import { useBaseRoute } from "../hooks/use-base-route";
 
 export default function App({ Component, pageProps }: AppProps) {
   const { isBaseRoute, currentBasePath } = useBaseRoute();
