+++ src/pages/work/[[...slug]].tsx
@@ -4,9 +4,9 @@ import { Work } from "@/sanity/types";
 import { useRouter } from "next/router";
 import { MediaRenderer } from "@/components/media-renderer";
 import { getWorkFirstMedia } from "@/sanity/lib/media";
-import ProjectSelector from "@/pages/work/components/selector";
-import ProjectPreview from "@/pages/work/components/preview";
-import ProjectCaseStudy from "@/pages/work/components/case-study";
+import ProjectSelector from "@/components/project/selector";
+import ProjectPreview from "@/components/project/preview";
+import ProjectCaseStudy from "@/components/project/case-study";
 import { worksApi } from "@/sanity/lib/work";
 import { Cross2Icon } from "@radix-ui/react-icons";
 import { GetStaticPaths, GetStaticProps } from "next";
